Refactored si_dynamic_t to si_array_t
I think it makes it more clear that this type is for handling more than one dynamically allocated value of the same size on the heap by changing the struct name from 'dynamic' to 'array'. Non-dynamic arrays could use this array struct although it does lose the type hint and could be accidentally free()'d.